Turn DisplayServer into an abstract class
Move Xorg specific code to XorgDisplayServer, this will allow us to
support multiple display server.
