+++ src/daemon/Seat.cpp
@@ -1,4 +1,5 @@
 /***************************************************************************
+* Copyright (c) 2014 Pier Luigi Fiorini <pierluigi.fiorini@gmail.com>
 * Copyright (c) 2013 Abdurrahman AVCI <abdurrahmanavci@gmail.com>
 *
 * This program is free software; you can redistribute it and/or modify
@@ -22,6 +23,7 @@
 #include "Configuration.h"
 #include "DaemonApp.h"
 #include "Display.h"
+#include "XorgDisplayServer.h"
 
 #include <QDebug>
 #include <QFile>
@@ -53,7 +55,10 @@ namespace SDDM {
         if (displayId == -1) {
             // find unused display
             displayId = findUnused(0, [&](const int number) {
-                return m_displayIds.contains(number) || QFile(QString("/tmp/.X%1-lock").arg(number)).exists();
+                bool alreadyExists = m_displayIds.contains(number);
+                if (!alreadyExists)
+                    alreadyExists = XorgDisplayServer::displayExists(number);
+                return alreadyExists;
             });
 
             // find unused terminal
@@ -69,7 +74,7 @@ namespace SDDM {
         m_terminalIds << terminalId;
 
         // log message
-        qDebug() << "Adding new display :" << displayId << " on vt" << terminalId << "...";
+        qDebug() << "Adding new display" << displayId << "on vt" << terminalId << "...";
 
         // create a new display
         Display *display = new Display(displayId, terminalId, this);
@@ -85,7 +90,7 @@ namespace SDDM {
     }
 
     void Seat::removeDisplay(int displayId) {
-        qDebug() << "Removing display :" << displayId << "...";
+        qDebug() << "Removing display" << displayId << "...";
 
         // display object
         Display *display = nullptr;
